Separate messageProcessor, make consumerInstance contructors private and reunite batch consumer with other consumer instance code
