+++ consumer/consumer_test.go
@@ -7,6 +7,7 @@ import (
 	"testing"
 
 	log "github.com/Financial-Times/go-logger/v2"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestConsume(t *testing.T) {
@@ -47,6 +48,21 @@ func TestConsume(t *testing.T) {
 				test.expMsgs, test.expErr, test.expCons, actMsgs, actErr, test.consumer.consumer)
 		}
 	}
+}
+
+func TestBatchConsumer(t *testing.T) {
+	consumer := &consumerInstance{
+		config:   QueueConfig{},
+		queue:    defaultTestQueueCaller{},
+		consumer: consInstTest, processor: batchedMessageProcessor{func(m []Message) {
+			assert.Equal(t, msgsTest, m)
+		}},
+		logger: log.NewUPPLogger("Test", "FATAL"),
+	}
+
+	msgs, err := consumer.consume()
+	assert.Nil(t, err)
+	assert.Equal(t, msgsTest, msgs)
 }
 
 func TestConsumeAndHandleMessagesRecoversFromPanic(t *testing.T) {
